--- a/taichi/program/async_utils.h
+++ b/taichi/program/async_utils.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <atomic>
-#include <optional>
 #include <unordered_map>
 #include <unordered_set>
 #include <variant>